Decouple background from container for opacity
