+++ src/panels/lovelace/hui-root.ts
@@ -78,6 +78,7 @@ import type { Lovelace } from "./types";
 import "./views/hui-view";
 import "./views/hui-view-container";
 import type { HUIView } from "./views/hui-view";
+import "./views/hui-background";
 
 @customElement("hui-root")
 class HUIRoot extends LitElement {
@@ -469,11 +470,11 @@ class HUIRoot extends LitElement {
         </div>
         <hui-view-container
           .hass=${this.hass}
-          .background=${background}
           .theme=${curViewConfig?.theme}
           id="view"
           @ll-rebuild=${this._debouncedConfigChanged}
         >
+          <hui-background .background=${background}> </hui-background>
         </hui-view-container>
       </div>
     `;
